[SC-36560] Implemented logic to destroy the client only when a user disconnects, preventing destruction on every error.

+++ src/index.js
@@ -18,15 +18,20 @@ export function runWhatsappClient (reauth = false) {
 
   run().catch(async(err) => {
     console.error("Error recieved from client: ", err)
+    const isUserDisconnect = err.message.includes("Execution context was destroyed, most likely because of a navigation.");
 
-    try {
-      await client.destroy();
-      console.log("WhatsApp client is disconnected.");
-    } catch (destroyErr) {
-      console.error("Error destroying client:", destroyErr);
+    if (isUserDisconnect) {
+      try {
+        await client.destroy();
+        console.log("WhatsApp client is disconnected.");
+        channel.speak({ reauthenticate: true});
+        runWhatsappClient(true);
+      } catch (destroyErr) {
+        console.error("Error destroying client:", destroyErr);
+      }
+    } else {
+      console.log("Error not related to user disconnect, skipping destroy.");
     }
-    channel.speak({ reauthenticate: true});
-    runWhatsappClient(true);
   })
 }
